tests: refactor some tests after the inclusion of state.py
