+++ tests/cli/test_status.py
@@ -17,15 +17,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import copy
 import io
-import subprocess
 import unittest
 import yaml
 
 from contextlib import redirect_stdout
-from unittest.mock import patch, call, mock_open
-from netplan.cli.commands.status import NetplanStatus, Interface
+from unittest.mock import patch
+from netplan.cli.commands.status import NetplanStatus
+from netplan.cli.state import Interface, SystemConfigState
 from tests.test_utils import call_cli
 
 
@@ -41,22 +40,6 @@ DNS_SEARCH = [(5, 'search.domain', False), (2, 'search.domain', False)]
 FAKE_DEV = {'ifindex': 42, 'ifname': 'fakedev0', 'flags': [], 'operstate': 'DOWN'}
 
 
-class resolve1_ipc_mock():
-    def get_object(self, _foo, _bar):
-        return {}  # dbus Object
-
-
-class resolve1_iface_mock():
-    def __init__(self, _foo, _bar):
-        pass  # dbus Interface
-
-    def GetAll(self, _):
-        return {
-            'DNS': DNS_ADDRESSES,
-            'Domains': DNS_SEARCH,
-            }
-
-
 class TestStatus(unittest.TestCase):
     '''Test netplan status'''
 
@@ -70,184 +53,8 @@ class TestStatus(unittest.TestCase):
     def _get_itf(self, ifname):
         return next((itf for itf in yaml.safe_load(IPROUTE2) if itf['ifname'] == ifname), None)
 
-    @patch('subprocess.check_output')
-    def test_query_iproute2(self, mock):
-        mock.return_value = IPROUTE2
-        status = NetplanStatus()
-        res = status.query_iproute2()
-        mock.assert_called_with(['ip', '-d', '-j', 'addr'], text=True)
-        self.assertEqual(len(res), 6)
-        self.assertListEqual([itf.get('ifname') for itf in res],
-                             ['lo', 'enp0s31f6', 'wlan0', 'wg0', 'wwan0', 'tun0'])
-
-    @patch('subprocess.check_output')
-    def test_query_iproute2_fail(self, mock):
-        mock.side_effect = subprocess.CalledProcessError(1, '', 'ERR')
-        status = NetplanStatus()
-        with self.assertLogs() as cm:
-            res = status.query_iproute2()
-            mock.assert_called_with(['ip', '-d', '-j', 'addr'], text=True)
-            self.assertIsNone(res)
-            self.assertIn('CRITICAL:root:Cannot query iproute2 interface data:', cm.output[0])
-
-    @patch('subprocess.check_output')
-    def test_query_networkd(self, mock):
-        mock.return_value = NETWORKD
-        status = NetplanStatus()
-        res = status.query_networkd()
-        mock.assert_called_with(['networkctl', '--json=short'], text=True)
-        self.assertEqual(len(res), 6)
-        self.assertListEqual([itf.get('Name') for itf in res],
-                             ['lo', 'enp0s31f6', 'wlan0', 'wg0', 'wwan0', 'tun0'])
-
-    @patch('subprocess.check_output')
-    def test_query_networkd_fail(self, mock):
-        mock.side_effect = subprocess.CalledProcessError(1, '', 'ERR')
-        status = NetplanStatus()
-        with self.assertLogs() as cm:
-            res = status.query_networkd()
-            mock.assert_called_with(['networkctl', '--json=short'], text=True)
-            self.assertIsNone(res)
-            self.assertIn('CRITICAL:root:Cannot query networkd interface data:', cm.output[0])
-
-    @patch('subprocess.check_output')
-    def test_query_nm(self, mock):
-        mock.return_value = NMCLI
-        status = NetplanStatus()
-        res = status.query_nm()
-        mock.assert_called_with(['nmcli', '-t', '-f',
-                                 'DEVICE,NAME,UUID,FILENAME,TYPE,AUTOCONNECT',
-                                 'con', 'show'], text=True)
-        self.assertEqual(len(res), 1)
-        self.assertListEqual([itf.get('device') for itf in res], ['wlan0'])
-
-    @patch('subprocess.check_output')
-    def test_query_nm_fail(self, mock):
-        mock.side_effect = subprocess.CalledProcessError(1, '', 'ERR')
-        status = NetplanStatus()
-        with self.assertLogs(level='DEBUG') as cm:
-            res = status.query_nm()
-            mock.assert_called_with(['nmcli', '-t', '-f',
-                                     'DEVICE,NAME,UUID,FILENAME,TYPE,AUTOCONNECT',
-                                     'con', 'show'], text=True)
-            self.assertIsNone(res)
-            self.assertIn('DEBUG:root:Cannot query NetworkManager interface data:', cm.output[0])
-
-    @patch('subprocess.check_output')
-    def test_query_routes(self, mock):
-        mock.side_effect = [ROUTE4, ROUTE6]
-        status = NetplanStatus()
-        res4, res6 = status.query_routes()
-        mock.assert_has_calls([
-            call(['ip', '-d', '-j', 'route'], text=True),
-            call(['ip', '-d', '-j', '-6', 'route'], text=True),
-            ])
-        self.assertEqual(len(res4), 6)
-        self.assertListEqual([route.get('dev') for route in res4],
-                             ['enp0s31f6', 'wlan0', 'wg0', 'enp0s31f6', 'wlan0', 'enp0s31f6'])
-        self.assertEqual(len(res6), 10)
-        self.assertListEqual([route.get('dev') for route in res6],
-                             ['lo', 'enp0s31f6', 'wlan0', 'enp0s31f6', 'wlan0',
-                              'tun0', 'enp0s31f6', 'wlan0', 'enp0s31f6', 'wlan0'])
-
-    @patch('subprocess.check_output')
-    def test_query_routes_fail(self, mock):
-        mock.side_effect = subprocess.CalledProcessError(1, '', 'ERR')
-        status = NetplanStatus()
-        with self.assertLogs(level='DEBUG') as cm:
-            res4, res6 = status.query_routes()
-            mock.assert_called_with(['ip', '-d', '-j', 'route'], text=True)
-            self.assertIsNone(res4)
-            self.assertIsNone(res6)
-            self.assertIn('DEBUG:root:Cannot query iproute2 route data:', cm.output[0])
-
-    @patch('dbus.Interface')
-    @patch('dbus.SystemBus')
-    def test_query_resolved(self, mock_ipc, mock_iface):
-        mock_ipc.return_value = resolve1_ipc_mock()
-        mock_iface.return_value = resolve1_iface_mock('foo', 'bar')
-        status = NetplanStatus()
-        addresses, search = status.query_resolved()
-        self.assertEqual(len(addresses), 4)
-        self.assertListEqual([addr[0] for addr in addresses],
-                             [5, 5, 2, 2])  # interface index
-        self.assertEqual(len(search), 2)
-        self.assertListEqual([s[1] for s in search],
-                             ['search.domain', 'search.domain'])
-
-    @patch('dbus.SystemBus')
-    def test_query_resolved_fail(self, mock):
-        mock.return_value = resolve1_ipc_mock()
-        mock.side_effect = Exception(1, '', 'ERR')
-        status = NetplanStatus()
-        with self.assertLogs(level='DEBUG') as cm:
-            addresses, search = status.query_resolved()
-            self.assertIsNone(addresses)
-            self.assertIsNone(search)
-            self.assertIn('DEBUG:root:Cannot query resolved DNS data:', cm.output[0])
-
-    def test_query_resolvconf(self):
-        status = NetplanStatus()
-        with patch('builtins.open', mock_open(read_data='''\
-nameserver 1.1.1.1
-nameserver 8.8.8.8
-options edns0 trust-ad
-search   some.domain
-search search.domain  another.one
-''')):
-            res = status.resolvconf_json()
-            print(res)
-            self.assertListEqual(res.get('addresses'), ['1.1.1.1', '8.8.8.8'])
-            self.assertListEqual(res.get('search'), ['search.domain', 'another.one'])
-            self.assertEqual(res.get('mode'), None)
-
-    def test_query_resolvconf_stub(self):
-        status = NetplanStatus()
-        with patch('builtins.open', mock_open(read_data='\
-# This is /run/systemd/resolve/stub-resolv.conf managed by man:systemd-resolved(8).')):
-            res = status.resolvconf_json()
-            self.assertEqual(res.get('mode'), 'stub')
-
-    def test_query_resolvconf_compat(self):
-        status = NetplanStatus()
-        with patch('builtins.open', mock_open(read_data='\
-# This is /run/systemd/resolve/resolv.conf managed by man:systemd-resolved(8).')):
-            res = status.resolvconf_json()
-            self.assertEqual(res.get('mode'), 'compat')
-
-    def test_query_resolvconf_fail(self):
-        status = NetplanStatus()
-        with self.assertLogs() as cm:
-            with patch('builtins.open', mock_open(read_data='')) as mock_file:
-                mock_file.side_effect = Exception(1, '', 'ERR')
-                status.resolvconf_json()
-                self.assertIn('WARNING:root:Cannot parse /etc/resolv.conf:', cm.output[0])
-
-    def test_query_online_state_online(self):
-        status = NetplanStatus()
-        dev = copy.deepcopy(FAKE_DEV)
-        dev['addr_info'] = [{
-            'local': '192.168.0.100',
-            'prefixlen': 24,
-        }]
-        dev['flags'].append('UP')
-        dev['operstate'] = 'UP'
-        routes = [{
-            'dst': 'default',
-            'gateway': '192.168.0.1',
-            'dev': dev['ifname'],
-        }]
-        dns = [(FAKE_DEV['ifindex'], 2, DNS_IP4)]
-        res = status.query_online_state([Interface(dev, [], [], (dns, None), (routes, None))])
-        self.assertTrue(res)
-
-    def test_query_online_state_offline(self):
-        status = NetplanStatus()
-        res = status.query_online_state([Interface(FAKE_DEV, [])])
-        self.assertFalse(res)
-
-    @patch('netplan.cli.commands.status.Interface.query_nm_ssid')
-    @patch('netplan.cli.commands.status.Interface.query_networkctl')
+    @patch('netplan.cli.state.Interface.query_nm_ssid')
+    @patch('netplan.cli.state.Interface.query_networkctl')
     def test_pretty_print(self, networkctl_mock, nm_ssid_mock):
         SSID = 'MYCON'
         nm_ssid_mock.return_value = SSID
@@ -256,11 +63,10 @@ search search.domain  another.one
             '''Activation Policy: manual
             WiFi access point: {} (b4:fb:e4:75:c6:21)'''.format(SSID)
 
-        status = NetplanStatus()
-        nd = status.process_networkd(NETWORKD)
-        nm = status.process_nm(NMCLI)
+        nd = SystemConfigState.process_networkd(NETWORKD)
+        nm = SystemConfigState.process_nm(NMCLI)
         dns = (DNS_ADDRESSES, DNS_SEARCH)
-        routes = (status.process_generic(ROUTE4), status.process_generic(ROUTE6))
+        routes = (SystemConfigState.process_generic(ROUTE4), SystemConfigState.process_generic(ROUTE6))
         fakeroute = {'type': 'local', 'dst': '10.0.0.0/16', 'gateway': '10.0.0.1', 'dev': FAKE_DEV['ifname']}
 
         interfaces = [
@@ -282,6 +88,7 @@ search search.domain  another.one
             data[ifname] = obj
         f = io.StringIO()
         with redirect_stdout(f):
+            status = NetplanStatus()
             status.pretty_print(data, len(interfaces)+1, _console_width=130)
             out = f.getvalue()
             self.assertEqual(out, '''\
@@ -339,24 +146,24 @@ search search.domain  another.one
 ''')
 
     @patch('netplan.cli.utils.systemctl')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_iproute2')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_networkd')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_nm')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_routes')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_resolved')
-    @patch('netplan.cli.commands.status.NetplanStatus.resolvconf_json')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_online_state')
+    @patch('netplan.cli.state.SystemConfigState.query_iproute2')
+    @patch('netplan.cli.state.SystemConfigState.query_networkd')
+    @patch('netplan.cli.state.SystemConfigState.query_nm')
+    @patch('netplan.cli.state.SystemConfigState.query_routes')
+    @patch('netplan.cli.state.SystemConfigState.query_resolved')
+    @patch('netplan.cli.state.SystemConfigState.resolvconf_json')
+    @patch('netplan.cli.state.SystemConfigState.query_online_state')
     def test_call_cli(self, online_mock, resolvconf_mock, rd_mock, routes_mock, nm_mock, networkd_mock, iproute2_mock,
                       systemctl_mock):
-        status = NetplanStatus()
         systemctl_mock.return_value = None
         iproute2_mock.return_value = [FAKE_DEV]
-        networkd_mock.return_value = status.process_networkd(NETWORKD)
         nm_mock.return_value = []
         routes_mock.return_value = (None, None)
         rd_mock.return_value = (None, None)
         resolvconf_mock.return_value = {'addresses': [], 'search': [], 'mode': None}
         online_mock.return_value = False
+        state = SystemConfigState()
+        networkd_mock.return_value = state.process_networkd(NETWORKD)
         out = self._call(['-a'])
         self.assertEqual(out.strip(), '''\
 Online state: offline
@@ -364,8 +171,8 @@ Online state: offline
 ● 42: fakedev0 other DOWN (unmanaged)''')
 
     @patch('netplan.cli.utils.systemctl')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_iproute2')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_networkd')
+    @patch('netplan.cli.state.SystemConfigState.query_iproute2')
+    @patch('netplan.cli.state.SystemConfigState.query_networkd')
     def test_fail_cli(self, networkd_mock, iproute2_mock, systemctl_mock):
         systemctl_mock.return_value = None
         iproute2_mock.return_value = [FAKE_DEV]
@@ -374,24 +181,24 @@ Online state: offline
             self._call([])
 
     @patch('netplan.cli.utils.systemctl')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_iproute2')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_networkd')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_nm')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_routes')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_resolved')
-    @patch('netplan.cli.commands.status.NetplanStatus.resolvconf_json')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_online_state')
+    @patch('netplan.cli.state.SystemConfigState.query_iproute2')
+    @patch('netplan.cli.state.SystemConfigState.query_networkd')
+    @patch('netplan.cli.state.SystemConfigState.query_nm')
+    @patch('netplan.cli.state.SystemConfigState.query_routes')
+    @patch('netplan.cli.state.SystemConfigState.query_resolved')
+    @patch('netplan.cli.state.SystemConfigState.resolvconf_json')
+    @patch('netplan.cli.state.SystemConfigState.query_online_state')
     def test_call_cli_ifname(self, online_mock, resolvconf_mock, rd_mock, routes_mock, nm_mock, networkd_mock, iproute2_mock,
                              systemctl_mock):
-        status = NetplanStatus()
         systemctl_mock.return_value = None
         iproute2_mock.return_value = [FAKE_DEV, self._get_itf('wlan0')]
-        networkd_mock.return_value = status.process_networkd(NETWORKD)
         nm_mock.return_value = []
         routes_mock.return_value = (None, None)
         rd_mock.return_value = (None, None)
         resolvconf_mock.return_value = {'addresses': [], 'search': [], 'mode': None}
         online_mock.return_value = False
+        state = SystemConfigState()
+        networkd_mock.return_value = state.process_networkd(NETWORKD)
         out = self._call([FAKE_DEV['ifname']])
         self.assertEqual(out.strip(), '''\
 Online state: offline
@@ -401,70 +208,70 @@ Online state: offline
 1 inactive interfaces hidden. Use "--all" to show all.''')
 
     @patch('netplan.cli.utils.systemctl')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_iproute2')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_networkd')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_nm')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_routes')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_resolved')
-    @patch('netplan.cli.commands.status.NetplanStatus.resolvconf_json')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_online_state')
+    @patch('netplan.cli.state.SystemConfigState.query_iproute2')
+    @patch('netplan.cli.state.SystemConfigState.query_networkd')
+    @patch('netplan.cli.state.SystemConfigState.query_nm')
+    @patch('netplan.cli.state.SystemConfigState.query_routes')
+    @patch('netplan.cli.state.SystemConfigState.query_resolved')
+    @patch('netplan.cli.state.SystemConfigState.resolvconf_json')
+    @patch('netplan.cli.state.SystemConfigState.query_online_state')
     def test_fail_cli_ifname(self, online_mock, resolvconf_mock, rd_mock, routes_mock, nm_mock, networkd_mock, iproute2_mock,
                              systemctl_mock):
-        status = NetplanStatus()
         systemctl_mock.return_value = None
         iproute2_mock.return_value = [FAKE_DEV, self._get_itf('wlan0')]
-        networkd_mock.return_value = status.process_networkd(NETWORKD)
         nm_mock.return_value = []
         routes_mock.return_value = (None, None)
         rd_mock.return_value = (None, None)
         resolvconf_mock.return_value = {'addresses': [], 'search': [], 'mode': None}
         online_mock.return_value = False
+        state = SystemConfigState()
+        networkd_mock.return_value = state.process_networkd(NETWORKD)
         with self.assertRaises(SystemExit):
             self._call(['notaninteface0'])
 
     @patch('netplan.cli.utils.systemctl')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_iproute2')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_networkd')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_nm')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_routes')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_resolved')
-    @patch('netplan.cli.commands.status.NetplanStatus.resolvconf_json')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_online_state')
+    @patch('netplan.cli.state.SystemConfigState.query_iproute2')
+    @patch('netplan.cli.state.SystemConfigState.query_networkd')
+    @patch('netplan.cli.state.SystemConfigState.query_nm')
+    @patch('netplan.cli.state.SystemConfigState.query_routes')
+    @patch('netplan.cli.state.SystemConfigState.query_resolved')
+    @patch('netplan.cli.state.SystemConfigState.resolvconf_json')
+    @patch('netplan.cli.state.SystemConfigState.query_online_state')
     def test_call_cli_json(self, online_mock, resolvconf_mock, rd_mock, routes_mock, nm_mock, networkd_mock, iproute2_mock,
                            systemctl_mock):
-        status = NetplanStatus()
         systemctl_mock.return_value = None
         iproute2_mock.return_value = [FAKE_DEV]
-        networkd_mock.return_value = status.process_networkd(NETWORKD)
         nm_mock.return_value = []
         routes_mock.return_value = (None, None)
         rd_mock.return_value = (None, None)
         resolvconf_mock.return_value = {'addresses': [], 'search': [], 'mode': None}
         online_mock.return_value = False
+        state = SystemConfigState()
+        networkd_mock.return_value = state.process_networkd(NETWORKD)
         out = self._call(['-a', '--format=json'])
         self.assertEqual(out, '''{\
 "netplan-global-state": {"online": false, "nameservers": {"addresses": [], "search": [], "mode": null}}, \
 "fakedev0": {"index": 42, "adminstate": "DOWN", "operstate": "DOWN"}}\n''')
 
     @patch('netplan.cli.utils.systemctl')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_iproute2')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_networkd')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_nm')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_routes')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_resolved')
-    @patch('netplan.cli.commands.status.NetplanStatus.resolvconf_json')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_online_state')
+    @patch('netplan.cli.state.SystemConfigState.query_iproute2')
+    @patch('netplan.cli.state.SystemConfigState.query_networkd')
+    @patch('netplan.cli.state.SystemConfigState.query_nm')
+    @patch('netplan.cli.state.SystemConfigState.query_routes')
+    @patch('netplan.cli.state.SystemConfigState.query_resolved')
+    @patch('netplan.cli.state.SystemConfigState.resolvconf_json')
+    @patch('netplan.cli.state.SystemConfigState.query_online_state')
     def test_call_cli_yaml(self, online_mock, resolvconf_mock, rd_mock, routes_mock, nm_mock, networkd_mock, iproute2_mock,
                            systemctl_mock):
-        status = NetplanStatus()
         systemctl_mock.return_value = None
         iproute2_mock.return_value = [FAKE_DEV]
-        networkd_mock.return_value = status.process_networkd(NETWORKD)
         nm_mock.return_value = []
         routes_mock.return_value = (None, None)
         rd_mock.return_value = (None, None)
         resolvconf_mock.return_value = {'addresses': [], 'search': [], 'mode': None}
         online_mock.return_value = False
+        state = SystemConfigState()
+        networkd_mock.return_value = state.process_networkd(NETWORKD)
         out = self._call(['-a', '--format=yaml'])
         self.assertEqual(out.strip(), '''\
 fakedev0:
@@ -478,33 +285,33 @@ netplan-global-state:
     search: []
   online: false'''.strip())
 
-    @patch('netplan.cli.commands.status.NetplanStatus.query_iproute2')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_networkd')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_nm')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_routes')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_resolved')
-    @patch('netplan.cli.commands.status.NetplanStatus.resolvconf_json')
-    @patch('netplan.cli.commands.status.NetplanStatus.query_online_state')
+    @patch('netplan.cli.state.SystemConfigState.query_iproute2')
+    @patch('netplan.cli.state.SystemConfigState.query_networkd')
+    @patch('netplan.cli.state.SystemConfigState.query_nm')
+    @patch('netplan.cli.state.SystemConfigState.query_routes')
+    @patch('netplan.cli.state.SystemConfigState.query_resolved')
+    @patch('netplan.cli.state.SystemConfigState.resolvconf_json')
+    @patch('netplan.cli.state.SystemConfigState.query_online_state')
     @patch('netplan.cli.utils.systemctl_is_active')
     @patch('netplan.cli.utils.systemctl')
     def test_call_cli_no_networkd(self, systemctl_mock, is_active_mock,
                                   online_mock, resolvconf_mock, rd_mock,
                                   routes_mock, nm_mock, networkd_mock,
                                   iproute2_mock):
-        status = NetplanStatus()
         iproute2_mock.return_value = [FAKE_DEV]
-        networkd_mock.return_value = status.process_networkd(NETWORKD)
         nm_mock.return_value = []
         routes_mock.return_value = (None, None)
         rd_mock.return_value = (None, None)
         resolvconf_mock.return_value = {'addresses': [], 'search': [], 'mode': None}
         online_mock.return_value = False
         is_active_mock.return_value = False
+        state = SystemConfigState()
+        networkd_mock.return_value = state.process_networkd(NETWORKD)
         with self.assertLogs(level='DEBUG') as cm:
             self._call([])
             self.assertIn('DEBUG:root:systemd-networkd.service is not active. Starting...',
                           cm.output[0])
-        systemctl_mock.assert_called_once_with('start', ['systemd-networkd.service'], True)
+        systemctl_mock.assert_called_with('start', ['systemd-networkd.service'], True)
 
     @patch('netplan.cli.utils.systemctl_is_active')
     @patch('netplan.cli.utils.systemctl_is_masked')
@@ -515,141 +322,3 @@ netplan-global-state:
             self._call([])
         self.assertEqual(1, e.exception.code)
         self.assertIn('systemd-networkd.service is masked', cm.output[0])
-
-
-class TestInterface(unittest.TestCase):
-    '''Test netplan status' Interface class'''
-
-    @patch('subprocess.check_output')
-    def test_query_nm_ssid(self, mock):
-        mock.return_value = ' MYSSID '  # added some whitespace to strip()
-        con = 'SOME_CONNECTION_ID'
-        itf = Interface(FAKE_DEV, [])
-        res = itf.query_nm_ssid(con)
-        mock.assert_called_with(['nmcli', '--get-values', '802-11-wireless.ssid',
-                                 'con', 'show', 'id', con],
-                                text=True)
-        self.assertEqual(res, 'MYSSID')
-
-    @patch('subprocess.check_output')
-    def test_query_nm_ssid_fail(self, mock):
-        mock.side_effect = subprocess.CalledProcessError(1, '', 'ERR')
-        con = 'SOME_CONNECTION_ID'
-        itf = Interface(FAKE_DEV, [])
-        with self.assertLogs() as cm:
-            res = itf.query_nm_ssid(con)
-            mock.assert_called_with(['nmcli', '--get-values', '802-11-wireless.ssid',
-                                     'con', 'show', 'id', con],
-                                    text=True)
-            self.assertIsNone(res)
-            self.assertIn('WARNING:root:Cannot query NetworkManager SSID for {}:'.format(con), cm.output[0])
-
-    @patch('subprocess.check_output')
-    def test_query_networkctl(self, mock):
-        mock.return_value = 'DOES NOT MATTER'
-        dev = 'fakedev0'
-        itf = Interface(FAKE_DEV, [])
-        res = itf.query_networkctl(dev)
-        mock.assert_called_with(['networkctl', 'status', '--', dev], text=True)
-        self.assertEqual(res, mock.return_value)
-
-    @patch('subprocess.check_output')
-    def test_query_networkctl_fail(self, mock):
-        mock.side_effect = subprocess.CalledProcessError(1, '', 'ERR')
-        dev = 'fakedev0'
-        itf = Interface(FAKE_DEV, [])
-        with self.assertLogs() as cm:
-            res = itf.query_networkctl(dev)
-            mock.assert_called_with(['networkctl', 'status', '--', dev], text=True)
-            self.assertIsNone(res)
-            self.assertIn('WARNING:root:Cannot query networkctl for {}:'.format(dev), cm.output[0])
-
-    @patch('netplan.cli.commands.status.Interface.query_nm_ssid')
-    @patch('netplan.cli.commands.status.Interface.query_networkctl')
-    def test_json_nm_wlan0(self, networkctl_mock, nm_ssid_mock):
-        SSID = 'MYCON'
-        nm_ssid_mock.return_value = SSID
-        # networkctl mock output reduced to relevant lines
-        networkctl_mock.return_value = \
-            'WiFi access point: {} (b4:fb:e4:75:c6:21)'.format(SSID)
-
-        status = NetplanStatus()
-        data = next((itf for itf in yaml.safe_load(IPROUTE2) if itf['ifindex'] == 5), {})
-        nd = status.process_networkd(NETWORKD)
-        nm = status.process_nm(NMCLI)
-        dns = (DNS_ADDRESSES, DNS_SEARCH)
-        routes = (status.process_generic(ROUTE4), status.process_generic(ROUTE6))
-
-        itf = Interface(data, nd, nm, dns, routes)
-        self.assertTrue(itf.up)
-        self.assertFalse(itf.down)
-        ifname, json = itf.json()
-        self.assertEqual(ifname, 'wlan0')
-        self.assertEqual(json.get('index'), 5)
-        self.assertEqual(json.get('macaddress'), '1c:4d:70:e4:e4:0e')
-        self.assertEqual(json.get('type'), 'wifi')
-        self.assertEqual(json.get('ssid'), 'MYCON')
-        self.assertEqual(json.get('backend'), 'NetworkManager')
-        self.assertEqual(json.get('id'), 'NM-b6b7a21d-186e-45e1-b3a6-636da1735563')
-        self.assertEqual(json.get('vendor'), 'Intel Corporation')
-        self.assertEqual(json.get('adminstate'), 'UP')
-        self.assertEqual(json.get('operstate'), 'UP')
-        self.assertEqual(len(json.get('addresses')), 4)
-        self.assertEqual(len(json.get('dns_addresses')), 2)
-        self.assertEqual(len(json.get('dns_search')), 1)
-        self.assertEqual(len(json.get('routes')), 6)
-
-    @patch('netplan.cli.commands.status.Interface.query_networkctl')
-    def test_json_nd_enp0s31f6(self, networkctl_mock):
-        # networkctl mock output reduced to relevant lines
-        networkctl_mock.return_value = 'Activation Policy: manual'
-
-        status = NetplanStatus()
-        data = next((itf for itf in yaml.safe_load(IPROUTE2) if itf['ifindex'] == 2), {})
-        nd = status.process_networkd(NETWORKD)
-        nm = status.process_nm(NMCLI)
-        dns = (DNS_ADDRESSES, DNS_SEARCH)
-        routes = (status.process_generic(ROUTE4), status.process_generic(ROUTE6))
-
-        itf = Interface(data, nd, nm, dns, routes)
-        self.assertTrue(itf.up)
-        self.assertFalse(itf.down)
-        ifname, json = itf.json()
-        self.assertEqual(ifname, 'enp0s31f6')
-        self.assertEqual(json.get('index'), 2)
-        self.assertEqual(json.get('macaddress'), '54:e1:ad:5f:24:b4')
-        self.assertEqual(json.get('type'), 'ethernet')
-        self.assertEqual(json.get('backend'), 'networkd')
-        self.assertEqual(json.get('id'), 'enp0s31f6')
-        self.assertEqual(json.get('vendor'), 'Intel Corporation')
-        self.assertEqual(json.get('adminstate'), 'UP')
-        self.assertEqual(json.get('operstate'), 'UP')
-        self.assertEqual(json.get('activation_mode'), 'manual')
-        self.assertEqual(len(json.get('addresses')), 3)
-        _, meta = list(json.get('addresses')[0].items())[0]  # get first (any only) address
-        self.assertIn('dhcp', meta.get('flags'))
-        self.assertEqual(len(json.get('dns_addresses')), 2)
-        self.assertEqual(len(json.get('dns_search')), 1)
-        self.assertEqual(len(json.get('routes')), 7)
-
-    def test_json_nd_tunnel(self):
-        status = NetplanStatus()
-        data = next((itf for itf in yaml.safe_load(IPROUTE2) if itf['ifindex'] == 41), {})
-        nd = status.process_networkd(NETWORKD)
-
-        itf = Interface(data, nd, [], (None, None), (None, None))
-        ifname, json = itf.json()
-        self.assertEqual(ifname, 'wg0')
-        self.assertEqual(json.get('index'), 41)
-        self.assertEqual(json.get('type'), 'tunnel')
-        self.assertEqual(json.get('backend'), 'networkd')
-        self.assertEqual(json.get('tunnel_mode'), 'wireguard')
-
-    def test_json_no_type_id_backend(self):
-        itf = Interface(FAKE_DEV, [], [], (None, None), (None, None))
-        ifname, json = itf.json()
-        self.assertEqual(ifname, 'fakedev0')
-        self.assertEqual(json.get('index'), 42)
-        self.assertNotIn('type', json)
-        self.assertNotIn('id', json)
-        self.assertNotIn('backend', json)
